requirements.txt
numpy>=1.21
pandas>=1.3
matplotlib>=3.5
PyYAML>=6.0
requests>=2.28
datasets>=2.0
transformers>=4.30
tiktoken>=0.5